requirements.txt
dash
plotly
langchain
langchain-openai
chromadb
openai
python-dotenv
pydantic
